52 - Add ownership to listing

+++ app/Http/Controllers/ListingController.php
@@ -48,11 +48,11 @@ class ListingController extends Controller
             'description' => 'required',
         ]);
 
-        if($request->hasFile('logo')) {
+        if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-        // $formFields['user_id'] = auth()->id();
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
@@ -84,10 +84,10 @@ class ListingController extends Controller
      */
     public function update(Request $request, Listing $listing)
     {
-        // if ($listing->user_id != auth()->id()) {
-        //     abort(403, 'Unauthorized Action');
-        // }
-        
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -98,7 +98,7 @@ class ListingController extends Controller
             'description' => 'required',
         ]);
 
-        if($request->hasFile('logo')) {
+        if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
@@ -113,11 +113,11 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
-        // if($listing->user_id != auth()->id()) {
-        //     abort(403, 'Unauthorized Action');
-        // }
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorised Action');
+        }
         
-        if($listing->logo && Storage::disk('public')->exists($listing->logo)) {
+        if ($listing->logo && Storage::disk('public')->exists($listing->logo)) {
             Storage::disk('public')->delete($listing->logo);
         }
         $listing->delete();
